(stablehlo.partition_id): add op handling and migrate to collectives.cc
- Added a new `collectives` op module and registered it in the build and handler table.
- Moved `stablehlo.all_reduce` handling out of `reduction.cc` into `collectives.cc`.
- Added `stablehlo.partition_id` support as a single-partition fallback that returns scalar zero and rejects multi-partition modules.
- Kept `stablehlo.all_reduce` as a single-device identity fallback, while still rejecting multi-device replica groups.
- Updated the unsupported-test skip pattern so `all_reduce` and `partition_id` are no longer treated as unsupported collectives.
- Added collectives test configs for single-device `lax.pmean` and `lax.axis_index`, and wired them into the main op test suite.
- Moved the existing all-reduce test out of reduction configs into the new collectives configs.
