+++ CMakeLists.txt
@@ -136,6 +136,7 @@ set(PJRT_SOURCES
     src/pjrt_plugin/ops/slice.cc
     src/pjrt_plugin/ops/gather_scatter.cc
     src/pjrt_plugin/ops/reduction.cc
+    src/pjrt_plugin/ops/collectives.cc
     src/pjrt_plugin/ops/linalg.cc
     src/pjrt_plugin/ops/control_flow.cc
     src/pjrt_plugin/ops/sort_fft_complex.cc
