+++ tests/configs/__init__.py
@@ -1,5 +1,6 @@
 from .benchmark import make_benchmark_op_configs
 from .binary import make_binary_op_configs
+from .collectives import make_collectives_op_configs
 from .control_flow import make_control_flow_op_configs
 from .conv import make_conv_op_configs
 from .conversion import make_conversion_op_configs
@@ -24,6 +25,7 @@ __all__ = [
     "OperationTestConfig",
     "make_benchmark_op_configs",
     "make_binary_op_configs",
+    "make_collectives_op_configs",
     "make_control_flow_op_configs",
     "make_conv_op_configs",
     "make_conversion_op_configs",
