+++ tests/configs/reduction.py
@@ -1,23 +1,7 @@
-import jax
 from jax import lax, random
 from jax import numpy as jnp
 
 from .util import OperationTestConfig, complex_standard_normal
-
-
-def _all_reduce_pmean(x):
-    """Single-device pmean, lowering to stablehlo.all_reduce (issue #201).
-
-    pmap ignores jax.default_device, so pin it to the harness's current default
-    device; otherwise the collective always runs on the default (MPS) backend and
-    the CPU comparison pass fails its device-placement assertion.
-    """
-    # getattr: jax.config exposes jax_default_device at runtime but pyright's
-    # stubs don't declare it.
-    dev = getattr(jax.config, "jax_default_device", None) or jax.devices()[0]
-    return jax.pmap(
-        lambda y: lax.pmean(y, axis_name="i"), axis_name="i", devices=[dev]
-    )(x)
 
 
 def make_reduction_op_configs():
@@ -98,17 +82,6 @@ def make_reduction_op_configs():
             lambda x: lax.reduce_window(x, 0.0, lax.add, (), (), "valid"),
             lambda key: random.normal(key, ()),
             name="reduce_window_scalar",
-        )
-
-    # stablehlo.all_reduce: single-device identity fallback (issue #201). pmap
-    # over one device emits all_reduce with a 1x1 replica_groups; the handler
-    # forwards operand -> result. The pmean VJP is itself an all_reduce, so the
-    # default grad path exercises the handler on the backward pass too.
-    with OperationTestConfig.module_name("reduction-real"):
-        yield OperationTestConfig(
-            _all_reduce_pmean,
-            lambda key: random.normal(key, (1, 4)),
-            name="all_reduce-pmean-identity",
         )
 
     # Pooling operations (lower to stablehlo.reduce_window with pooling pattern).
